Move EventJson to its own file
